fix the MacOS compilation error (#18)
* fix the MacOS compilation error

* change the types to uint64_t

+++ src/simplex.cc
@@ -148,7 +148,7 @@ void SimplexDecoder::init_ilp() {
   highs->setOptionValue("output_flag", config.verbose);
 }
 
-void SimplexDecoder::decode_to_errors(const std::vector<size_t>& detections) {
+void SimplexDecoder::decode_to_errors(const std::vector<uint64_t>& detections) {
   predicted_errors_buffer.clear();
   // Adjust the constraints for the detection events
   for (size_t d : detections) {
@@ -345,7 +345,7 @@ common::ObservablesMask SimplexDecoder::mask_from_errors(
 }
 
 common::ObservablesMask SimplexDecoder::decode(
-    const std::vector<size_t>& detections) {
+    const std::vector<uint64_t>& detections) {
   decode_to_errors(detections);
   return mask_from_errors(predicted_errors_buffer);
 }
